+++ src/modules/hip/hip_tensor_statistical_operations.hpp
@@ -27,5 +27,6 @@ SOFTWARE.
 #include "kernel/tensor_sum.hpp"
 #include "kernel/tensor_min.hpp"
 #include "kernel/tensor_max.hpp"
+#include "kernel/normalize.hpp"
 
 #endif // HIP_TENSOR_STATISTICAL_OPERATIONS_HPP